refactor(kicad-studio): extract component search view
